Update Loper to support NSSecureCoding
iOS 14 is more strict about NSCoding and NSSecureCoding protocols.  This PR updates Loper to play nice with NSSecureCoding.

One thing I found was the the `hardReset` method was erroring out.  It's only called in the tests, but the error was with the DispatchQueue deadlocking on `self.close()`.  The note said "Mutex is recursive", but the tests said otherwise.

Because NSSecureCoding enforces typecasting during decode, we could remove the expected conditional in `EncodedObject.swift`

Some auto formatting too.